requirements.txt
django>=5.2.5
djangorestframework>=3.15
django-cors-headers>=4.4
django-filter>=23.5
pandas>=2.2
openpyxl>=3.1
xlrd>=2.0
gunicorn>=22.0
whitenoise>=6.6.0
dj-database-url>=2.2.0
python-decouple>=3.8
psycopg2-binary>=2.9
django-import-export>=4.1.1
tablib[xls,xlsx]>=3.5.0
google-genai
